Add a Zig "wrapper" for `sim86_shared.h`
This wrapper implementation differs from others
as it does not utilize any of the precompiled
libraries. It leverages Zig's interop with C
(see links below), to build the C++ shared
library and then interface with it via the
`sim86_shared.h` C/C++ header file.

The `sim86.zig` file includes Zig `type-aliases`
for the natively supported `c.*` types that are
imported from the Zig-compiled library. It also
adds small wrappers around strings, returning an
error for an unrecognized Op, and Zig enum
support for `OperandType`. It also follows the
"traditional" Zig naming conventions.

This commit also adds a simple test file that
demonstrates the correctness of this
implementation.

See `README.md` for more information.

Links:

- [C Interop](https://ziglang.org/learn/overview/#integration-with-c-libraries-without-ffibindings)